backend/requirements.txt
fastapi
uvicorn
pillow
ultralytics
python-multipart
keras
tensorflow